# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator",
        "description": "API endpoints to manage articles",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "List articles",
                "description": "Get a paginated list of articles.",
                "operationId": "20f46a4be21edb0179e4df7acaa45c18",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword to filter articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "provider",
                        "in": "query",
                        "description": "provider to filter articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category to filter articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter articles for selected date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "next page",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleCollection"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid query parameters"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/single/{articleId}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a single article by articleId",
                "operationId": "3f0c93ddb6f950c610eec77004e808b7",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ab9ecf70-2fa2-4817-9564-34e5a5a5b98d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Single article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "User registration",
                "operationId": "235fcf4a18b375b0a91c567f3e272dce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid data provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in an existing user",
                "description": "User login",
                "operationId": "41fe9a5872f93c5a0f8458ac3b93ea98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out the current user",
                "description": "Log out the current user",
                "operationId": "c3c84aa0043a5e1fbad0a8fe659badb3",
                "responses": {
                    "200": {
                        "description": "User logged out successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserPreferenceResource": {
                "properties": {
                    "uuid": {
                        "description": "The unique identifier for the preference",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "providers": {
                        "description": "List of providers",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "authors": {
                        "description": "List of authors",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categories": {
                        "description": "List of categories",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "uuid": {
                        "description": "The unique identifier for the user",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "description": "The name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "The email of the user",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "token": {
                        "description": "The user's authentication token",
                        "type": "string",
                        "example": "your-jwt-token",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "The user account creation date",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "updated_at": {
                        "description": "The last update date of the user account",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "preference": {
                        "$ref": "#/components/schemas/UserPreferenceResource"
                    }
                },
                "type": "object"
            },
            "ArticleResource": {
                "required": [
                    "uuid",
                    "title",
                    "description",
                    "source",
                    "published_at",
                    "url",
                    "category",
                    "image_url",
                    "provider",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "title": {
                        "type": "string",
                        "example": "Article Title"
                    },
                    "description": {
                        "type": "string",
                        "example": "This is a description of the article."
                    },
                    "source": {
                        "type": "string",
                        "example": "Provider Name"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://example.com/article"
                    },
                    "category": {
                        "type": "string",
                        "example": "Category Name"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg"
                    },
                    "provider": {
                        "type": "string",
                        "example": "Provider Name"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-26T12:34:56"
                    }
                },
                "type": "object"
            },
            "ArticleCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/ArticleResource"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ArticleResource"
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token to access the API endpoints",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}